# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Telecom Management - API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/cupi/user/delete/{username}": {
            "delete": {
                "tags": [
                    "Management - UnityConnection"
                ],
                "summary": "Delete User",
                "description": "This deletes the user by username",
                "operationId": "deleteuser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "username",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/cupi/usertemplate/create": {
            "post": {
                "tags": [
                    "Management - UnityConnection"
                ],
                "summary": "Create User Template",
                "description": "",
                "operationId": "createUserTemplate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sitecode",
                        "in": "formData",
                        "description": "New Template Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "New Template Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "language",
                        "in": "formData",
                        "description": "Template to Copy",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "timezone",
                        "in": "formData",
                        "description": "Timezone",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "operator",
                        "in": "formData",
                        "description": "Timezone",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/cupi/usertemplate/site": {
            "post": {
                "tags": [
                    "Management - UnityConnection"
                ],
                "summary": "Create Employee and Partner Templates for Site",
                "description": "",
                "operationId": "createUserTemplate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sitecode",
                        "in": "formData",
                        "description": "New Template Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "language",
                        "in": "formData",
                        "description": "Language",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "english",
                            "french"
                        ]
                    },
                    {
                        "name": "timezone",
                        "in": "formData",
                        "description": "Timezone",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "Alaska-12",
                            "Arizona-12",
                            "Central-12",
                            "Eastern-12",
                            "Hawaii-12",
                            "Mountain-12",
                            "Pacific-12"
                        ]
                    },
                    {
                        "name": "operator",
                        "in": "formData",
                        "description": "Timezone",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/cupi/usertemplate/deletebyname/{name}": {
            "delete": {
                "tags": [
                    "Management - UnityConnection"
                ],
                "summary": "Delete UserTemplate",
                "description": "This deletes the user template by name",
                "operationId": "delete_usertemplate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Template Alias",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/cupi/usertemplates/listusertemplatesbysite/{sitecode}": {
            "get": {
                "tags": [
                    "Management - UnityConnection"
                ],
                "summary": "Get UM External Service for User Object",
                "description": "",
                "operationId": "getUserTemplateBySite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sitecode",
                        "in": "path",
                        "description": "Sitecode",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
        "/api/cupi/usertemplates/list_call_handlers": {
            "get": {
                "tags": [
                    "Management - UnityConnection"
                ],
                "summary": "List Call Handlers",
                "description": "",
                "operationId": "list_call_handlers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
        "/api/cupi/timezones": {
            "get": {
                "tags": [
                    "Management - UnityConnection"
                ],
                "summary": "List Timezones",
                "description": "",
                "operationId": "ListTimeZones",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
    },
...
}